backend/knowledge-base-api/requirements.txt
FastAPI
uvicorn
sqlalchemy
pydantic
milvus
python-dotenv
alembic
pytest
pytest-asyncio
httpx